+++ NET/LibUA/Types.cs
@@ -6909,7 +6909,7 @@ namespace LibUA
 
             public uint ChannelID { get; set; }
             public uint TokenID { get; set; }
-            public double TokenLifetime { get; set; }
+            public uint TokenLifetime { get; set; }
             public DateTimeOffset TokenCreatedAt { get; set; }
 
             public uint? PrevChannelID { get; set; }
